Updated Thread fields according to domain model

+++ src/main/java/com/endregaswarriors/meddit/repositories/internal/ThreadRepository.java
@@ -3,6 +3,6 @@ package com.endregaswarriors.meddit.repositories.internal;
 import com.endregaswarriors.meddit.models.database.Thread;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ThreadRepository extends CrudRepository<Thread, Integer> {
+public interface ThreadRepository extends CrudRepository<Thread, Long> {
 
 }
